Move struct to internal header (#239)
decoder_pkey needs EVP_KEYMGMT structure, so move it from
evp_local.h to a new internal/evp.h header file
